+++ copyright.txt
@@ -1,0 +1,7 @@
+{pkg.name}
+ ---
+{pkg.description}
+@version: v{pkg.version}
+@author: {pkg.author}
+@link: {pkg.homepage}
+@license: {pkg.license}
