+++ src/Terms.ts
@@ -11,13 +11,7 @@ import { WpApiParent } from './Parent';
 
 import { WpApiLoader } from './Loaders';
 
-export interface IWpApiTerms {
-  getList(taxonomiesType: string, options?: RequestOptionsArgs): Observable<Response>;
-  get(taxonomiesType: string, termId: number, options?: RequestOptionsArgs): Observable<Response>;
-  create(taxonomiesType: string, body: any, options?: RequestOptionsArgs): Observable<Response>;
-  update(taxonomiesType: string, termId: number, body: any, options?: RequestOptionsArgs): Observable<Response>;
-  delete(taxonomiesType: string, termId: number, options?: RequestOptionsArgs): Observable<Response>;
-}
+import { IWpApiTerms } from './interfaces';
 
 const defaultTaxoType = 'categories';
